requirements.txt
streamlit>=1.28.0
PyPDF2>=4.0.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langchain-community>=0.0.1
langchain-core>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
deep-translator>=1.11.0
openai>=1.3.0
pdfplumber>=0.10.0
rank-bm25>=0.2.2
chromadb>=0.4.0
huggingface-hub>=0.17.0
pymupdf>=1.24.0



streamlit>=1.35.0
groq>=0.9.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10